requirements.txt
streamlit
python-dotenv
google-generativeai
datetime
requests
pymupdf
langchain
qdrant-client
beautifulsoup4